Create top-level Form design component.
