+++ apps/spotlight/src/pages/ud105-evicition-form.astro
@@ -1,9 +1,10 @@
 ---
-import { FormView } from '@atj/design/src/FormManager/FormView';
+import { Form } from '@atj/design';
+
 import formData from '../htmlParser/ud105-form-field-output.json';
 import ContentLayout from '../layouts/ContentLayout.astro';
 ---
 
 <ContentLayout title="10x Access to Justice Spotlight">
-  <FormView prompt={formData} />
+  <Form prompt={formData} />
 </ContentLayout>
