+++ packages/design/src/FormManager/FormView/index.tsx
@@ -1,10 +1,7 @@
 import React from 'react';
-import { FormProvider, useForm } from 'react-hook-form';
 
-import { type Prompt, createFormContext, createPrompt } from '@atj/forms';
 import { type FormService } from '@atj/form-service';
-
-import { PromptSegment } from '../prompts';
+import Form from '../../Form';
 
 // Assuming this is the structure of your JSON data
 export interface Field {
@@ -28,12 +25,10 @@ export const FormViewById = ({
   if (!result.success) {
     return 'null form retrieved from storage';
   }
-  const context = createFormContext(result.data);
-  const prompt = createPrompt(context);
 
   return (
-    <FormView
-      prompt={prompt}
+    <Form
+      form={result.data}
       onSubmit={async data => {
         const submission = await formService.submitForm(formId, data);
         if (submission.success) {
@@ -45,97 +40,6 @@ export const FormViewById = ({
         }
       }}
     />
-  );
-};
-
-export const FormView = ({
-  prompt,
-  onSubmit,
-}: {
-  prompt: Prompt;
-  onSubmit?: (data: Record<string, string>) => void;
-}) => {
-  const formMethods = useForm<Record<string, string>>({});
-  return (
-    <FormProvider {...formMethods}>
-      <form
-        onSubmit={formMethods.handleSubmit(async data => {
-          if (onSubmit) {
-            console.log('Submitting form...');
-            onSubmit(data);
-          } else {
-            console.warn('Skipping form submission...');
-          }
-        })}
-      >
-        <fieldset className="usa-fieldset">
-          {prompt.map((promptPart, index) => (
-            <PromptSegment key={index} promptPart={promptPart}></PromptSegment>
-          ))}
-          {/* Add submit button or other controls as needed */}
-        </fieldset>
-        <ButtonBar />
-      </form>
-    </FormProvider>
-  );
-};
-
-/*
-export const FormFieldsetUnwired = ({ fields }: { fields: Field[] }) => {
-  return (
-    <fieldset className="usa-fieldset">
-      <legend className="usa-legend usa-legend--large">
-        UD 105 - Unlawful Detainer Form
-      </legend>
-      {fields.map(field => {
-        // Use 'tag' for 'select' and 'textarea', 'type' for others
-        const fieldType =
-          field.tag === 'select' ||
-          field.tag === 'textarea' ||
-          field.tag === 'p' ||
-          field.tag === 'h2' ||
-          field.tag === 'h3' ||
-          field.tag === 'ul'
-            ? field.tag
-            : field.type;
-
-        switch (fieldType) {
-          case 'text':
-            return <TextField key={field.id} field={field} />;
-          case 'boolean':
-            return <BooleanField key={field.id} field={field} />;
-          case 'checkbox':
-            return <CheckBoxField key={field.id} field={field} />;
-          case 'select':
-            return <SelectField key={field.id} field={field} />;
-          case 'radio':
-            return <RadioField key={field.id} field={field} />;
-          case 'date':
-            return <DateField key={field.id} field={field} />;
-          case 'textarea':
-            return <TextareaField key={field.id} field={field} />;
-          case 'p':
-            return <ParagraphBlock key={field.id} field={field} />;
-          case 'h2':
-            return <Header3Block key={field.id} field={field} />;
-          case 'h3':
-            return <Header3Block key={field.id} field={field} />;
-          case 'ul':
-            return <UnorderedList key={field.id} field={field} />;
-          default:
-            return null;
-        }
-      })}
-    </fieldset>
-  );
-};
-*/
-
-const ButtonBar = () => {
-  return (
-    <div>
-      <button className="usa-button">Submit</button>
-    </div>
   );
 };
