+++ packages/design/src/index.ts
@@ -1,3 +1,4 @@
 import '@uswds/uswds';
 
+export { default as Form } from './Form';
 export { default as FormManager } from './FormManager';
