+++ packages/types/custom.d.ts
@@ -1,5 +1,9 @@
-type Success<T> = { success: true; data: T };
-type VoidSuccess = { success: true };
-type Failure<E> = { success: false; error: E };
-type Result<T, E = string> = Success<T> | Failure<E>;
-type VoidResult<E = string> = VoidSuccess | Failure<E>;
+declare global {
+  type Success<T> = { success: true; data: T };
+  type VoidSuccess = { success: true };
+  type Failure<E> = { success: false; error: E };
+  type Result<T, E = string> = Success<T> | Failure<E>;
+  type VoidResult<E = string> = VoidSuccess | Failure<E>;
+}
+
+export {};
